+++ src/app/app.json.ts
@@ -24,8 +24,24 @@ export const examples = [
                         'test2',
                         {
                             tag: 'div',
-                            attrs: { id: 'test-subdiv', class: 'foobar' },
-                            body: 'test3',
+                            attrs: {
+                                id: 'test-subdiv',
+                                class: 'foobar',
+                                style: 'all: initial;',
+                                lang: 'ja',
+                                'data-test1': 'value1',
+                                'data-test2': 'value2',
+                                'data-test3': 'value3',
+                            },
+                            body:
+                                `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem ` +
+                                `Ipsum has been the industry's standard dummy text ever since the 1500s, ` +
+                                `when an unknown printer took a galley of type and scrambled it to make a type ` +
+                                `specimen book. It has survived not only five centuries, but also the leap into ` +
+                                `electronic typesetting, remaining essentially unchanged. It was popularised in the ` +
+                                `1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more ` +
+                                `recently with desktop publishing software like Aldus PageMaker including versions ` +
+                                `of Lorem Ipsum.`,
                         },
                     ],
                 },